app/requirements.txt
beautifulsoup4
requests
pandas
psycopg2-binary
flask
